pyproject.toml
[tool.poetry]
name = "class-big-data"
version = "0.1.0"
description = ""
authors = ["Steven Santacruz Garcia <steven.santacruz@finkargo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyspark = "^3.5.4"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
